requirements.txt
fastapi
uvicorn
websockets
torch
numpy
pandas
httpx
python-dotenv